@@ -65,33 +58,38 @@ to interact with the data coming from the FSW.
     ####
     classifiers=[
         # complete classifier list: http://pypi.python.org/pypi?%3Aaction=list_classifiers
-        'Development Status :: 5 - Production/Stable',
-        'Intended Audience :: Developers',
-        'License :: OSI Approved :: Apache 2.0',
-        'Operating System :: Unix',
-        'Operating System :: POSIX',
-        'Programming Language :: Python',
-        'Programming Language :: Python :: 2.7',
-        'Programming Language :: Python :: 3',
-        'Programming Language :: Python :: 3.4',
-        'Programming Language :: Python :: 3.5',
-        'Programming Language :: Python :: 3.6',
-        'Programming Language :: Python :: 3.7',
-        'Programming Language :: Python :: Implementation :: CPython',
-        'Programming Language :: Python :: Implementation :: PyPy',
+        "Development Status :: 5 - Production/Stable",
+        "Intended Audience :: Developers",
+        "License :: OSI Approved :: Apache 2.0",
+        "Operating System :: Unix",
+        "Operating System :: POSIX",
+        "Programming Language :: Python",
+        "Programming Language :: Python :: 3",
+        "Programming Language :: Python :: 3.5",
+        "Programming Language :: Python :: 3.6",
+        "Programming Language :: Python :: 3.7",
+        "Programming Language :: Python :: Implementation :: CPython",
+        "Programming Language :: Python :: Implementation :: PyPy",
     ],
-    # Requires Python 2.7, or 3.4+, and the 'six' Python package
-    python_requires='>=2.7, !=3.0.*, !=3.1.*, !=3.2.*, !=3.3.*',
-    install_requires=['six', "lxml", 'enum34;python_version < "3.4"', "Markdown", "pexpect", "pytest",
-                  'Cheetah3;python_version >= "3.0"', 'Cheetah;python_version < "3.0"'],
+    # Requires Python 3.5+, and the 'six' Python package
+    python_requires=">=3.5",
+    install_requires=[
+        "six",
+        "lxml",
+        'enum34;python_version < "3.4"',
+        "Markdown",
+        "pexpect",
+        "pytest",
+        'Cheetah3;python_version >= "3.0"',
+        'Cheetah;python_version < "3.0"',
+    ],
+    extras_require={"dev": ["black", "pylama", "pylint", "pre-commit"]},
     # Setup and test requirments, not needed by normal install
-    setup_requires=['pytest-runner'],
-    tests_require=['pytest'],
+    setup_requires=["pytest-runner"],
+    tests_require=["pytest"],
     # Create a set of executable entry-points for running directly from the package
     entry_points={
-        "console_scripts": [
-            "fprime-util = fprime.util.__main__:main"
-        ],
-        "gui_scripts": []
-    }
+        "console_scripts": ["fprime-util = fprime.util.__main__:main"],
+        "gui_scripts": [],
+    },
 )
