@@ -1,21 +1,16 @@
-'''
+"""
 Created on Dec 18, 2014
 
 @author: tcanham
 
-'''
-from __future__ import print_function
-from __future__ import absolute_import
-import struct
-from .type_exceptions import *
-from . import type_base
-from . import u32_type
-from . import string_type
-from . import enum_type
+"""
+import copy
 
-@type_base.serialize
-@type_base.deserialize
-class SerializableType(type_base.BaseType):
+from .type_base import BaseType, ValueType
+from .type_exceptions import NotInitializedException, TypeMismatchException
+
+
+class SerializableType(ValueType):
     """
     Representation of the Serializable type (comparable to the ANY type)
 
@@ -25,121 +20,124 @@ class SerializableType(type_base.BaseType):
     @param param: typename = "SomeTypeName" string
     To preserve member order, the member argument is a list of members and their types:
     @param param: mem_list = [ ("member",<ref to BaseType>, format string, description), ... ]
-                  OR mem_list = [ ("member",<ref to BaseType>, format string), ... ].
-                  The member descriptions can be None
+    OR mem_list = [ ("member",<ref to BaseType>, format string), ... ].
+    The member descriptions can be None
     """
-    def __init__(self, typename, mem_list = None):
+
+    def __init__(self, typename, mem_list=None):
         """
         Constructor
         """
-
-        if not type(typename) == type(str()):
-            raise TypeMismatchException(type(str()),type(typename))
-
+        super().__init__()
+        if not isinstance(typename, str):
+            raise TypeMismatchException(str, type(typename))
         self.__typename = typename
-        setattr(self, "mem_list", None)
+        # If the member list is defined, stamp in None for any missing descriptions
+        if mem_list:
+            new_mem_list = [
+                entry if len(entry) == 4 else (entry[0], entry[1], entry[2], None)
+                for entry in mem_list
+            ]
+        # Set the member list then set the value
+        self.mem_list = new_mem_list
 
-        if mem_list == None or len(mem_list) == 0:
-            return;
-
-        if not type(mem_list) == type(list()):
-            raise TypeMismatchException(type(list()),type(mem_list))
-
-        # Check if the member list contains a description or not
-        if len(mem_list[0]) == 3:
-            new_mem_list = []
-            for (memberName, memberVal, format_string) in mem_list:
-                new_mem_list.append((memberName, memberVal, format_string, None))
-
-            mem_list = new_mem_list
-
-        # scan the list to see if it has the correct types
-        for (memberName, memberVal, format_string, desc) in mem_list:
-            # member name should be a string
-            if not type(memberName) == type(str()):
-                raise TypeMismatchException(type(str()),type(memberName))
-            # member value should be a derived class of TypeBase
-            if not issubclass(type(memberVal), type_base.BaseType):
-                raise TypeMismatchException(type(type_base.BaseType),type(memberVal))
-            # format string should be string
-            if not type(format_string) == type(str()):
-                raise TypeMismatchException(type(str()),type(format_string))
-            # Description should be a string
-            if desc != None and not type(desc) == type(str()):
-                raise TypeMismatchException(type(str()),type(desc))
-
-        self.__mem_list = mem_list
-
-    def to_jsonable(self):
-        """
-        JSONable type
-        """
-        members = {}
-        for member in self.mem_list:
-            members[member[0]] = {"format": member[2], "description": member[3]}
-            members[member[0]].update(member[1].to_jsonable())
-        return members
+    def validate(self, val=None):
+        """ Validate this object including member list and values """
+        # Blank member list does not validate
+        if not self.mem_list:
+            return
+        elif not isinstance(self.mem_list, list):
+            raise TypeMismatchException(list, self.mem_list)
+        for member_name, member_val, format_string, description in self.mem_list:
+            # Check each of these members for correct types
+            if not isinstance(member_name, str):
+                raise TypeMismatchException(str, type(member_name))
+            elif not isinstance(member_val, BaseType):
+                raise TypeMismatchException(BaseType, type(member_val))
+            elif not isinstance(format_string, str):
+                raise TypeMismatchException(str, type(format_string))
+            elif description is not None and not isinstance(description, str):
+                raise TypeMismatchException(str, type(description))
+        # When a value is set and is not empty we need to set the member properties
+        if not val:
+            return
+        # If a value is supplied then check each value against the member list
+        for val_member, list_entry in zip(val, self.mem_list):
+            _, member_list_val, _, _ = list_entry
+            member_list_val.validate(
+                val_member
+            )  # Insure that the the val_member is consistent with the existing member
 
     @property
     def mem_list(self):
+        """ Gets the member list"""
         return self.__mem_list
 
-
     @mem_list.setter
-    def mem_list(self, ml):
+    def mem_list(self, mem_list):
         """
-        @todo: add arg type checking
+        Sets the member list and validates against the current value.
         """
-        self.__mem_list = ml
-
+        self.__mem_list = mem_list
+        if mem_list is None:
+            self.validate(self.mem_list)
 
     def serialize(self):
-        if self.mem_list == None:
+        """ Serializes the members of the serializable """
+        if self.mem_list is None:
             raise NotInitializedException(type(self))
-
-        # iterate through members and serialize each one
-        serStream = ""
-        for (memberName, memberVal, format_string, desc) in self.mem_list:
-            serStream += memberVal.serialize()
-
-        return serStream
-
+        return b"".join(
+            [member_val.serialize() for _, member_val, _, _ in self.mem_list]
+        )
 
     def deserialize(self, data, offset):
-        self.__val = []
-        for (memberName, memberVal, format_string, desc) in self.mem_list:
-            memberVal.deserialize(data, offset)
-            self.__val.append(memberVal.val)
-            offset += memberVal.getSize()
+        """ Deserialize the values of each of the members """
+        new_member_list = []
+        for entry1, member_val, entry3, entry4 in self.mem_list:
+            cloned = copy.copy(member_val)
+            cloned.deserialize(data, offset)
+            new_member_list.append((entry1, cloned, entry3, entry4))
+            offset += member_val.getSize()
+        self.mem_list = new_member_list
 
     @property
-    def val(self):
-        return self.__val
+    def val(self) -> dict:
+        """
+        The .val property typically returns the python-native type. This the python native type closes to a serializable
+        without generating full classes would be a dictionary (anonymous object). This returns such an object.
+
+        :return dictionary of member names to python values of member keys
+        """
+        return {
+            member_name: member_val.val
+            for member_name, member_val, _, _ in self.mem_list
+        }
+
+    @val.setter
+    def val(self, val: dict):
+        """
+        The .val property typically returns the python-native type. This the python native type closes to a serializable
+        without generating full classes would be a dictionary (anonymous object). This takes such an object and sets the
+        member val list from it.
+
+        :param val: dictionary containing python types to key names. This
+        """
+        values_list = [val[name] for name, _, _, _ in self.mem_list]
+        # Member list is the explicit store for storing these values
+        for val_member, list_entry in zip(values_list, self.mem_list):
+            _, member_list_val, _, _ = list_entry
+            member_list_val.val = val_member
 
     def getSize(self):
-        size = 0
-        for (memberName, memberVal, format_string, desc) in self.mem_list:
-            size += memberVal.getSize()
-        return size;
+        """ The size of a struct is the size of all the members """
+        return sum([mem_type.getSize() for _, mem_type, _, _ in self.mem_list])
 
-if __name__ == '__main__':
-    print("Serializable")
-    try:
-        i32Mem    = u32_type.U32Type(1000000)
-        stringMem = string_type.StringType("something to say")
-        members   = { "MEMB1":0 , "MEMB2":6, "MEMB3":9 }
-        enumMem   = enum_type.EnumType("SomeEnum",members,"MEMB3")
-        #print enumMem.val
-
-        memList = [ ("mem1",i32Mem),("mem2",stringMem),("mem3",enumMem)]
-
-        serType = SerializableType("ASerType", memList)
-
-        print("Value: %s" % repr(memList))
-        buff = serType.serialize()
-        type_base.showBytes(buff)
-        serType2 = SerializableType("ASerType",memList)
-        serType2.deserialize(buff, len(buff))
-        print("Deserialized: %s" % repr(serType2.mem_list))
-    except TypeException as e:
-        print("Exception: %s" % e.getMsg())
+    def to_jsonable(self):
+        """
+        JSONable type for a serializable
+        """
+        members = {}
+        for member_name, member_value, member_format, member_desc in self.mem_list:
+            members[member_name] = {"format": member_format, "description": member_desc}
+            members[member_name].update(member_value.to_jsonable())
+        return members
